+++ Application/Features/StudentCourseTable/Commands/Create/CreateStudentCourseCommand.cs
@@ -4,7 +4,7 @@ using MediatR;
 
 namespace Application.Features.StudentCourseTable.Commands.Create
 {
-    public class CreateStudentCourseCommand : StudentCourseDTO, IRequest<int>
+    public class CreateStudentCourseCommand : StudentCourseDTO, IRequest<long>
   {
     public CreateStudentCourseCommand()
     { }
@@ -17,7 +17,7 @@ namespace Application.Features.StudentCourseTable.Commands.Create
       CourseId = dto.CourseId;
 
     }
-    public class Handler : IRequestHandler<CreateStudentCourseCommand, int>
+    public class Handler : IRequestHandler<CreateStudentCourseCommand, long>
     {
       private readonly IApplicationDbContext _context;
       public Handler(IApplicationDbContext context)
@@ -25,7 +25,7 @@ namespace Application.Features.StudentCourseTable.Commands.Create
 
         _context = context;
       }
-      public async Task<int> Handle(CreateStudentCourseCommand request, CancellationToken cancellationToken)
+      public async Task<long> Handle(CreateStudentCourseCommand request, CancellationToken cancellationToken)
       {
         Domain.Entities.StudentCourse entity = new Domain.Entities.StudentCourse
         {
