+++ Application/Features/StudentCourseTable/Commands/Update/UpdateStudentCourseCommand.cs
@@ -5,7 +5,7 @@ using MediatR;
 
 namespace Application.Features.StudentCourseTable.Commands.Update
 {
-    public class UpdateStudentCourseCommand : StudentCourseDTO, IRequest<int>
+    public class UpdateStudentCourseCommand : StudentCourseDTO, IRequest<long>
   {
     public UpdateStudentCourseCommand()
     { }
@@ -16,7 +16,7 @@ namespace Application.Features.StudentCourseTable.Commands.Update
       Id = dto.Id;
 
     }
-    public class Handler : IRequestHandler<UpdateStudentCourseCommand, int>
+    public class Handler : IRequestHandler<UpdateStudentCourseCommand, long>
     {
       private readonly IApplicationDbContext _context;
       public Handler(IApplicationDbContext context)
@@ -24,7 +24,7 @@ namespace Application.Features.StudentCourseTable.Commands.Update
 
         _context = context;
       }
-      public async Task<int> Handle(UpdateStudentCourseCommand request, CancellationToken cancellationToken)
+      public async Task<long> Handle(UpdateStudentCourseCommand request, CancellationToken cancellationToken)
       {
         Domain.Entities.StudentCourse entity = new Domain.Entities.StudentCourse
         {
